Spinlock, atomic test-and-set, creating processes.

+++ include/string.h
@@ -5,7 +5,7 @@
 
 void* memcpy(void* dest, const void* src, uint32_t len);
 
-void* memset(void* dest, uint8_t val, uint32_t len);
+void* memset(void* dest, int val, uint32_t len);
 
 int strcmp(const char* str1, const char* str2);
